requirements.txt
accelerate==0.27.2
build==1.0.3
cachetools==5.3.3
Deprecated==1.2.14
einops==0.7.0
Flask==2.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
googleapis-common-protos==1.62.0
gptcache==0.1.43
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
hdbcli==2.19.21
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
importlib-metadata==6.11.0
jmespath==1.0.1
langchain==0.1.13
langchain-community==0.0.29
openai==1.12.0
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
pdfplumber==0.11.0
Werkzeug==2.2.2
python-dotenv
pymongo
langchain_mongodb
langchain_openai
pymupdf
langchainhub
docx2txt
python-docx
langchain_experimental
pandas
tabulate
matplotlib